backend/pyproject.toml
[project]
name = "studyforge-backend"
version = "0.1.0"
description = "StudyForge AI Study Workspace - Backend API"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",

    # Auth
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",

    # Validation & settings
    "pydantic>=2.7.0",
    "email-validator>=2.0.0",
    "pydantic-settings>=2.3.0",

    # Task queue
    "celery[redis]>=5.4.0",

    # Redis
    "redis>=5.0.0",

    # S3
    "boto3>=1.34.0",

    # PDF processing
    "pymupdf>=1.24.0",

    # AI
    "openai>=1.30.0",
    "anthropic>=0.49.0",

    # Utilities
    "httpx>=0.27.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
